add new signal preprocessor that uses Smoothing spline
The Smoothing spline is implemented by using the Univariate spline from the scipy library. The optimal smoothness parameter is determined by leave one of CV. Given the computation burden on the CV, the cache manager saves the optimal smoothness parameter for future use. The get_processed_data function is refactored to accommodate the new preprocessor.  Both signal preprocessors are working on the original data and if standardization was requested then the interpolated data is transformed.